Fix the physics of pioneer2dx
It turns out the reason of incorrect of physics is because of a plugin is used by the original pioneer2dx model. Our compiled version can correctly load the plugin because of a rpath setting, while the original gzserver cannot.

The physics looks reasonable after removing the plugin from the model.