Fix memory leak from pygazebo.World.step
Step() uses gazebo::runWorld(), which in turn uses World::RunLoop(). RunLoop() starts LogWorker()
every time. LogWorker will always store something in the buffer when it is started. But we/ don't
have a running LogRecord to collect all these data. This cause memory usage keep  increasing.

We create a LogRecord to solve this issue.